csv/requirements.txt
mysql-connector-python
pandas
configparser
flask_sqlalchemy
pymysql
